pyproject.toml
[tool.poetry]
name = "modernfiction"
version = "0.1.0"
description = ""
authors = ["John Suh <johnsuh94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
markitdown = "^0.0.1a3"
aiohttp = {extras = ["speedups"], version = "^3.11.11"}
pillow = "^11.0.0"
pypdf = "^5.1.0"
pdf2image = "^1.17.0"
python-dotenv = "^1.0.1"
retry = "^0.9.2"
ebooklib = "^0.18"
tqdm = "^4.67.1"
epubcheck = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
